Move pkg/api/ref to only user plugin/pkg/auth/authorizer/node
