+++ plugin/pkg/auth/authorizer/node/graph.go
@@ -19,9 +19,9 @@ package node
 import (
 	"sync"
 
-	pvutil "k8s.io/kubernetes/pkg/api/persistentvolume"
 	api "k8s.io/kubernetes/pkg/apis/core"
 	podutil "k8s.io/kubernetes/pkg/apis/core/helper/pod"
+	pvutil "k8s.io/kubernetes/plugin/pkg/auth/authorizer/node/persistentvolume"
 	"k8s.io/kubernetes/third_party/forked/gonum/graph"
 	"k8s.io/kubernetes/third_party/forked/gonum/graph/simple"
 )
